How to  Create UIWindow, Navigation Controller , Navigation Bar Using Xib
1.	Create an Empty User Interface and name it MainWindow.xib
2.	Add UIWindow to it from the Object Library to the canvas.
3.	Add NSObject from the Object library to the Placeholder/object Panel
(just below the UIWindow) and name it “AppDelegate”.
(This will show all the IBOutlet to the respective object in the xib)
4.	Now Add NavigationController from the ObjectLibrary just parallel to
the UIWindow in the canvas.
5.	Conntect the File Owner of the Xib to the UIApplication.
6.	Create two IBOutlet  in the AppDelegate
IBOutlet UIWindow *window;
IBOutlet UINavigationController *navigationController;

And remove the stuff written programmatically.

7.	In AppDelegate.h
Set UIResponder ⇒  NSObject.

This will show all the IBOutlet to the respective objects in the Xib.
8.	Set the File Owner Delegate as navigation Window.
9.	In MainWindow.Xib, go to window(object panel) set
                rootViewController  → navigationController
		window ⇒ AppDelegate
10.	Set the Delegate of navigation Controller as file owner’s
Navigation Controller ⇒ AppDelegate
rootViewController ⇒ window

11. Create a view using IBoutlet UIview *containerView and connect
            containterView to their respective IBOutlet   and add
button set IBAction and set push View Controller.

Set the Main Interface of the Application as MainWindow